Refactored code to create a domains package.
